requirements.txt
# Основные библиотеки для бота
python-binance==1.0.19
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Дополнительные библиотеки для работы с данными
aiohttp==3.9.1
websockets==12.0

# Логирование и отладка
coloredlogs==15.0

# Дополнительные зависимости для стабильной работы
asyncio-throttle==1.0.2
pytz==2023.3